feat: 베스트 독서 기록문 스크린 제작 (데스크탑 해상도) (#3)
