+++ src/components/composites/main/bestCommentary/desktop/BestCommentaryDesktop.styled.ts
@@ -1,0 +1,7 @@
+import tw from 'twin.macro';
+
+export const BestCommentarySection = tw.section`tablet:mb-[64px] mx-[64px] px-[24px] py-[16px] bg-white`;
+
+export const Title = tw.h2`tablet:t-body-m15 mb-[8px] desktop:d-body-m17`;
+
+export const BestCommentaryCardWrapper = tw.div`grid grid-cols-auto-fill-240 gap-x-[40px] gap-y-[20px]`;
