+++ src/components/composites/main/bestCommentary/desktop/BestCommentaryDesktop.tsx
@@ -1,7 +1,26 @@
 import React from 'react';
 
-const BestCommentaryDesktop = () => {
-  return <div>BestCommentaryDesktop</div>;
+import { BookRecordCard } from '@/components';
+import type { BookRecordCardType } from '@/types';
+import * as S from './BestCommentaryDesktop.styled';
+
+interface BestCommentaryDesktopProps {
+  books: BookRecordCardType[];
+}
+
+const BestCommentaryDesktop = ({ books }: BestCommentaryDesktopProps) => {
+  return (
+    <S.BestCommentarySection>
+      <header>
+        <S.Title>주간 기록 베스트 10!</S.Title>
+      </header>
+      <S.BestCommentaryCardWrapper>
+        {books.map((book) => (
+          <BookRecordCard key={book.id} linkToBaseUrl="book" {...book} />
+        ))}
+      </S.BestCommentaryCardWrapper>
+    </S.BestCommentarySection>
+  );
 };
 
 export default BestCommentaryDesktop;
